requirements.txt
python-telegram-bot==22.5
requests
python-dotenv
supabase
